pyproject.toml
[project]
name = "demo-agent-01"
version = "0.1.0"
description = "AI Agent Demo using Docker-based MCP for GitHub integration"
authors = [
    { name = "Yuval Reuveni", email = "yuval.reuveni@checkmarx.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["ai", "agent", "mcp", "github", "docker"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",

    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "strands-agents>=1.8.0",
    "strands-agents-tools>=0.2.7", 
    "mcp==1.13.1",
    "python-dotenv",
    "requests",
    "beautifulsoup4",
    "gitpython",
    "openai",
    "langfuse",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.20.0",
]

[project.scripts]
demo-agent = "demo_agent.cli:main"
github-agent = "agents.github_agent:main"
test-mcp = "test_docker_mcp:main"

[project.urls]
Homepage = "https://github.com/cx-yuval-reuveni/demo-agent-01"
Repository = "https://github.com/cx-yuval-reuveni/demo-agent-01"
Documentation = "https://github.com/cx-yuval-reuveni/demo-agent-01/blob/main/README.md"
Issues = "https://github.com/cx-yuval-reuveni/demo-agent-01/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_tools", "agents"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agent_tools", "agents"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--cov=agent_tools",
    "--cov=agents", 
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "strands.*",
    "langfuse.*", 
    "mcp.*",
    "beautifulsoup4.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["agent_tools", "agents"]
omit = [
    "*/tests/*",
    "*/test_*.py", 
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]